Redis: Extract TargetIndexer class
