+++ lib/redis/index.js
@@ -1,10 +1,10 @@
 'use strict'
 
-var Keys = require('./keys')
 var PromiseHelper = require('./promise-helper')
 var ListHelper = require('./list-helper')
 var SearchHelper = require('./search-helper')
 var AutocompleteIndexer = require('./autocomplete-indexer')
+var TargetIndexer = require('./target-indexer')
 
 module.exports = RedisClient
 
@@ -12,36 +12,6 @@ function RedisClient(clientImpl, config, getTimestamp) {
   this.impl = clientImpl
   this.rangeSize = (config || {}).REDIS_RANGE_SIZE || 25
   this.getTimestamp = getTimestamp || (() => new Date().getTime())
-}
-
-class TargetIndexer {
-  constructor(clientImpl) {
-    this.impl = clientImpl
-  }
-
-  addLink(link, target) {
-    return PromiseHelper.do(done => {
-      this.impl.lpush(Keys.targetIndex(target), link, done)
-    })
-  }
-
-  reindexLink(link, prevTarget, newTarget) {
-    return prevTarget === newTarget ? Promise.resolve() : Promise.all([
-      this.addLink(link, newTarget),
-      this.removeLink(link, prevTarget)
-    ])
-  }
-
-  removeLink(link, target) {
-    return new ListHelper(this.impl).removeItem(Keys.targetIndex(target), link)
-  }
-
-  getLinksToTarget(target) {
-    var getLinks = done => {
-      this.impl.lrange(Keys.targetIndex(target), 0, -1, done)
-    }
-    return PromiseHelper.do(getLinks).then(links => links.sort())
-  }
 }
 
 RedisClient.prototype.fetchLinkData = function(links) {
